FIFA-2022-Predictions
A machine learning model to predict the 2022 FIFA world cup champions. The model predicts the final score of each match from group stages to finals. 
This match score is based on the match history of the two teams, and the playing venue is considered. Home teams have a little edge in winning the match most of the time. 
If the predicted scores are tied, there will be a penalty shootout, and the winner will be selected randomly. 
The notebook contains three sections: the exploratory data analysis section, building ML models, and the FIFA 2022 World Cup section.